add event detail page

+++ backend/api/views.py
@@ -97,27 +97,28 @@ def create_event(request):
       return Response(serializer.data, status=status.HTTP_201_CREATED)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
-@permission_classes([IsAdminUser])
+@api_view(['GET', 'PATCH', 'DELETE'])
+@permission_classes([AllowAny]) 
 def event_detail(request, event_id):
-  event = get_object_or_404(Event, pk=event_id)
-  #view event
-  if request.method == 'GET':
-    serializer = EventSerializer(event)
-    return Response(serializer.data)
+    event = get_object_or_404(Event, pk=event_id)
+    if request.method == 'GET':
+        serializer = EventSerializer(event)
+        return Response(serializer.data)
+    elif request.method == 'PATCH':
+        if not request.user.is_staff:
+            return Response({'detail': 'Permission denied. Admin access required.'}, status=status.HTTP_403_FORBIDDEN)
+        
+        serializer = EventSerializer(event, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-  #update event
-  elif request.method == 'PATCH':
-    serializer = EventSerializer(event, data=request.data, partial=True)
-    if serializer.is_valid():
-      serializer.save()
-      return Response(serializer.data)
-    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-  #delete event
-  elif request.method == 'DELETE':
-    event.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    elif request.method == 'DELETE':
+        if not request.user.is_staff:
+            return Response({'detail': 'Permission denied. Admin access required.'}, status=status.HTTP_403_FORBIDDEN)
+        event.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 @api_view(["GET"])
 def view_all_skill(request):
